requirements.txt
Django>=4.2
mysqlclient
python-decouple
djangorestframework
